+++ lib/src/app.cpp
@@ -90,7 +90,7 @@ auto App::change_mount_point(std::string_view const directory) -> bool {
 
 auto App::get_features() const -> FeatureFlags {
 	auto ret = do_get_native_features();
-	if (get_render_device().validation_layers_enabled()) { ret.set(Feature::validation_layers); }
+	if (get_render_device().validation_layers_enabled()) { ret.set(Feature::eValidationLayers); }
 	return ret;
 }
 
@@ -118,7 +118,7 @@ void App::push_event(Event event) {
 	if (auto const* pointer_tap = std::get_if<PointerTap>(&event)) {
 		m_gesture_recognizer.on_tap(*pointer_tap);
 	} else if (auto const* key_input = std::get_if<KeyInput>(&event)) {
-		auto key_state = m_key_state.held_keys[static_cast<std::size_t>(key_input->key)];
+		auto key_state = m_key_state.held_keys[key_input->key];
 		switch (key_input->action) {
 		case Action::ePress: key_state = true; break;
 		case Action::eRelease: key_state = false; break;