+++ lib/docs/changelog.md
@@ -7,8 +7,8 @@
 - Added bave::Gamepad support, queried through bave::App::get_gamepad() and bave::App::get_gamepads().
   - Not yet supported on Android.
 - Added bave::KeyState and bave::App::get_key_state().
+- Added bave::EnumFlags.
 
 ### Changes
 
-- Replaced `bave::mod::*` with bave::Mod.
-- Refactored bave::ParticleEmitter::Modifier into an enum.
+- Unified all bit flags to be bave::EnumFlags.